requirements.txt
tensorflow-cpu==2.10.0
numpy==1.23.5
matplotlib==3.6.2
pillow==9.3.0 